+++ javascript/app-framework/docs/markdown/app-framework.plugincomponentprops.md
@@ -1,0 +1,18 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/app-framework](./app-framework.md) &gt; [PluginComponentProps](./app-framework.plugincomponentprops.md)
+
+## PluginComponentProps interface
+
+<b>Signature:</b>
+
+```typescript
+export interface PluginComponentProps<T extends Resource> 
+```
+
+## Properties
+
+|  Property | Modifiers | Type | Description |
+|  --- | --- | --- | --- |
+|  [resource](./app-framework.plugincomponentprops.resource.md) |  | T |  |
+
