+++ javascript/app-framework/docs/markdown/app-framework.plugincomponentprops.resource.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/app-framework](./app-framework.md) &gt; [PluginComponentProps](./app-framework.plugincomponentprops.md) &gt; [resource](./app-framework.plugincomponentprops.resource.md)
+
+## PluginComponentProps.resource property
+
+<b>Signature:</b>
+
+```typescript
+resource: T;
+```
