+++ javascript/app-framework/src/components/editors/DataTableEditor/DataTableEditor.tsx
@@ -2,39 +2,64 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
+import { ArqueroDetailsList, ArqueroTableHeader } from '@datashaper/react'
+import { ToolPanel } from '@essex/components'
+import { CommandBar } from '@fluentui/react'
 import { memo } from 'react'
-import { useHelpOnMount } from '../../../hooks/useHelpOnMount.js'
-import { CompoundEditor } from '../CompoundEditor/index.js'
-import { DataTableTextEditor } from '../DataTableTextEditor/index.js'
-import { ResourceSchemaEditor } from '../ResourceSchemaEditor/index.js'
 
+import { useDataTableSource } from '../../../hooks/index.js'
+import { DataTableConfig } from '../../DataTableConfig/DataTableConfig.js'
+import { useToolPanelExpandCollapse } from '../hooks.js'
+import {
+	useTableHeaderColors,
+	useTableHeaderStyles,
+	useToolPanelStyles,
+} from '../styles.js'
+import {
+	ConfigContainer,
+	Container,
+	DetailsListContainer,
+} from './DataTableEditor.styles.js'
 import type { DataTableEditorProps } from './DataTableEditor.types.js'
-import { TableEditor } from './TableEditor.js'
 
 export const DataTableEditor: React.FC<DataTableEditorProps> = memo(
 	function DataTableEditor({ resource }) {
-		useHelpOnMount('resources.datatable.index')
-		return <CompoundEditor resource={resource} editors={editors} />
+		const table = useDataTableSource(resource)
+		const tableHeaderColors = useTableHeaderColors()
+		const tableHeaderStyles = useTableHeaderStyles()
+		const toolPanelStyles = useToolPanelStyles()
+		const { collapsed, onToggleCollapsed, commandBar, iconProps } =
+			useToolPanelExpandCollapse('options-button', 'DataManagementSettings')
+		return table?.table == null ? null : (
+			<Container collapsed={collapsed}>
+				<DetailsListContainer>
+					<ArqueroTableHeader
+						background={tableHeaderColors.background}
+						styles={tableHeaderStyles}
+						farCommandBar={<CommandBar {...commandBar} />}
+						table={table.table}
+					/>
+					<ArqueroDetailsList
+						compact
+						sortable
+						showColumnBorders
+						isHeaderFixed
+						fill
+						metadata={table.metadata}
+						table={table?.table}
+					/>
+				</DetailsListContainer>
+				<ToolPanel
+					headerText={'Parser options'}
+					onDismiss={onToggleCollapsed}
+					headerIconProps={iconProps}
+					styles={toolPanelStyles}
+				>
+					<ConfigContainer>
+						<DataTableConfig resource={resource} />
+					</ConfigContainer>
+				</ToolPanel>
+			</Container>
+		)
 	},
 )
-
-const editors = [
-	{
-		key: 'data-table-text',
-		title: 'View and edit raw table text content',
-		iconName: 'PageData',
-		renderer: DataTableTextEditor,
-	},
-	{
-		key: 'data-table-json',
-		title: 'View and edit table JSON config',
-		iconName: 'Code',
-		renderer: ResourceSchemaEditor,
-	},
-	{
-		key: 'data-table-ux',
-		title: 'View and edit interactive table options',
-		iconName: 'PreviewLink',
-		renderer: TableEditor,
-	},
-]
