--- javascript/app-framework/src/components/editors/DataTableEditor/TableEditor.types.ts
@@ -1,9 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-import type { DataTable } from '@datashaper/workflow'
-
-export interface TableEditorProps {
-	resource: DataTable
-}
