+++ javascript/app-framework/src/profiles/DataTableProfile.ts
@@ -6,16 +6,17 @@ import content from '@datashaper/guidance'
 import type { DataTable } from '@datashaper/workflow'
 import { DataTableProfile as DataTableProfileBase } from '@datashaper/workflow'
 
-import { DataTableEditor, RawTableViewer } from '../components/editors/index.js'
+import { RawTableViewer } from '../components/editors/index.js'
 import type { GeneratedExtraRoutes, ProfilePlugin } from '../types.js'
 import { ResourceGroupType } from '../types.js'
+import { DataTableRenderer } from './renderers/DataTableRenderer.js'
 
 export class DataTableProfile
 	extends DataTableProfileBase
 	implements ProfilePlugin<DataTable>
 {
 	public readonly title = 'Datatable'
-	public readonly renderer = DataTableEditor
+	public readonly renderer = DataTableRenderer
 	public readonly iconName = 'PageData'
 	public readonly group = ResourceGroupType.Data
 	public readonly dataHandler = null
